refactor: Spot sealed class 적용
