--- src/main/kotlin/minesweeper/domain/MineSpot.kt
@@ -1,7 +1,0 @@
-package minesweeper.domain
-
-class MineSpot(coordinate: Coordinate) : Spot(coordinate) {
-    override fun isMine(): Boolean {
-        return true
-    }
-}
